+++ database/migrations/2024_07_31_133037_create_affiliate_payments_table.php
@@ -12,17 +12,19 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('affiliate_payments', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->string('phone');
-            $table->integer('amount');
-            $table->string('currency');
-            $table->string('payment_method');
-            $table->string('transaction_id');
-            $table->string('status')->default('Pending')->comment('Pending|Completed|Failed|Cancelled');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('affiliate_payments')) {
+            Schema::create('affiliate_payments', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->string('phone');
+                $table->integer('amount');
+                $table->string('currency');
+                $table->string('payment_method');
+                $table->string('transaction_id');
+                $table->string('status')->default('Pending')->comment('Pending|Completed|Failed|Cancelled');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
