+++ database/migrations/2024_08_07_145120_create_affiliate_withdraw_transactions_table.php
@@ -12,14 +12,16 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('affiliate_withdraw_transactions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('prev_balance')->default(0);
-            $table->string('withdraw_amount')->default(0);
-            $table->tinyInteger('status')->default(0)->comment('0=Pending|1=Approved|2=Rejected');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('affiliate_withdraw_transactions')) {
+            Schema::create('affiliate_withdraw_transactions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->integer('prev_balance')->default(0);
+                $table->string('withdraw_amount')->default(0);
+                $table->tinyInteger('status')->default(0)->comment('0=Pending|1=Approved|2=Rejected');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
