+++ database/migrations/2024_12_23_172939_create_merchant_payment_withdraws_table.php
@@ -12,14 +12,16 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('merchant_payment_withdraws', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId("user_id")->constrained("users")->onDelete('cascade');
-            $table->foreignId("store_id")->constrained("stores")->onDelete('cascade');
-            $table->string('withdraw_amount')->default(0);
-            $table->tinyInteger('status')->default(0)->comment('0=Pending|1=Approved|2=Completed|3=Rejected');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('merchant_payment_withdraws')) {
+            Schema::create('merchant_payment_withdraws', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId("user_id")->constrained("users")->onDelete('cascade');
+                $table->foreignId("store_id")->constrained("stores")->onDelete('cascade');
+                $table->string('withdraw_amount')->default(0);
+                $table->tinyInteger('status')->default(0)->comment('0=Pending|1=Approved|2=Completed|3=Rejected');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
