+++ database/migrations/2025_01_01_185001_create_superstaff_sales_commission_balances_table.php
@@ -12,25 +12,27 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('superstaff_sales_commission_balances', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId("user_id")->nullable()->constrained("users")->onDelete('cascade');
-            $table->unsignedBigInteger('staff_id');
-            $table->tinyInteger('isSetup')->default(0)->comment("0=No|1=Yes");
-            $table->tinyInteger('isNew')->default(0)->comment("0=No|1=Yes");
-            $table->tinyInteger('isRenew')->default(0)->comment("0=No|1=Yes");
-            $table->decimal('setup_commission')->default(0.00);
-            $table->decimal('new_commission')->default(0.00);
-            $table->decimal('renew_commission')->default(0.00);
-            $table->decimal('total_amount')->default(0.00);
-            $table->decimal('commission_amount')->default(0.00);
-            $table->decimal('dr')->default(0.00);
-            $table->decimal('cr')->default(0.00);
-            $table->decimal('balance')->default(0.00);
-            $table->timestamps();
-
-            $table->foreign('staff_id')->references('id')->on('superstaffs')->onDelete('cascade');
-        });
+        if (! Schema::hasTable('superstaff_sales_commission_balances')) {
+            Schema::create('superstaff_sales_commission_balances', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId("user_id")->nullable()->constrained("users")->onDelete('cascade');
+                $table->unsignedBigInteger('staff_id');
+                $table->tinyInteger('isSetup')->default(0)->comment("0=No|1=Yes");
+                $table->tinyInteger('isNew')->default(0)->comment("0=No|1=Yes");
+                $table->tinyInteger('isRenew')->default(0)->comment("0=No|1=Yes");
+                $table->decimal('setup_commission')->default(0.00);
+                $table->decimal('new_commission')->default(0.00);
+                $table->decimal('renew_commission')->default(0.00);
+                $table->decimal('total_amount')->default(0.00);
+                $table->decimal('commission_amount')->default(0.00);
+                $table->decimal('dr')->default(0.00);
+                $table->decimal('cr')->default(0.00);
+                $table->decimal('balance')->default(0.00);
+                $table->timestamps();
+    
+                $table->foreign('staff_id')->references('id')->on('superstaffs')->onDelete('cascade');
+            });
+        }
     }
 
     /**
