+++ database/migrations/2025_01_15_112341_create_admin_chat_supports_table.php
@@ -12,13 +12,15 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('admin_chat_supports', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId("store_id")->constrained("stores")->onDelete('cascade');
-            $table->integer("support")->default(0);
-            $table->dateTime("date")->nullable();
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('admin_chat_supports')) {
+            Schema::create('admin_chat_supports', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId("store_id")->constrained("stores")->onDelete('cascade');
+                $table->integer("support")->default(0);
+                $table->dateTime("date")->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
